+++ src/dse_polars/__init__.py
@@ -6,18 +6,22 @@ except PackageNotFoundError:
     # Package is not installed (e.g., running from a local script)
     __version__ = "unknown"
 
-from .dse import DSE, drop_subref
+from .dse import DSE
 from .texts import DSEPassages, ctsurn_contains
-from .images import CitableIIIFService, urn2info_url, info_url2urn, roi, strip_roi
+from .images import CitableIIIFService, urn2info_url, info_url2urn, roi, strip_roi, ptinrect, rois, urn2image_url
 from .urnutils import passagecomponent_re
 
 __all__ = ["DSE", 
+           "passagecomponent_re",
            "DSEPassages",
            "ctsurn_contains",
-           "drop_subref",  
+         
                  
            "CitableIIIFService",
            "urn2info_url",
            "info_url2urn",
            "roi","strip_roi",
-           "passagecomponent_re"]
+            "ptinrect",
+            "rois",
+            "urn2image_url"
+        ]